+++ haarpy/_utils.py
@@ -87,6 +87,7 @@ def _generate_matrices_with_row_sums(
     iterator of tuple[tuple[int, ...], ...]
         Yields all nonnegative integer matrix with prescribed row sum
     """
+
     def generate_compositions(total: int, length: int) -> Iterable[tuple[int, ...]]:
         "Generate all fixed length tuples of nonnegative integers summing to total"
         if length == 1:
@@ -120,6 +121,7 @@ def _vector_multinomial(
     int
         the product of the multinomial coefficients
     """
+
     def multinomial(total: int, parts: tuple[int, ...]) -> int:
         """Multinomial coefficient total! / prod parts_i!"""
         if sum(parts) != total:
@@ -181,7 +183,7 @@ def _matrix_to_sequence(
     col_index = tuple(
         i
         for row in power_matrix
-        for i, j in zip(tuple(range(len(power_matrix))), row)
+        for i, j in zip(tuple(range(len(power_matrix[0]))), row)
         for _ in range(j)
     )
 
@@ -224,5 +226,44 @@ def _sequence_to_matrix(
     for position, index_values in enumerate(zip(row_index_tuple, col_index_tuple)):
         for r, c in zip(*index_values):
             matrix_list[position][index_dict[r]][index_dict[c]] += 1
+
+    return [tuple(tuple(row) for row in matrix) for matrix in matrix_list]
+
+
+def _compressed_unitary_pair(
+    matrix_tuple: tuple[tuple[tuple[int, ...], tuple[tuple[int, ...]]]],
+) -> tuple[tuple[tuple[int, ...], tuple[tuple[int, ...]]]]:
+    """Simplifies a pair of power matrices for Haar unitary integral
+
+    Parameters
+    ----------
+    matrix_tuple: tuple[tuple[tuple[int, ...], tuple[tuple[int, ...]]]]
+        two power matrices to be compressed
+
+    Returns
+    -------
+    tuple[tuple[tuple[int, ...], tuple[tuple[int, ...]]]]
+        the simpliest power matrices yielding the same Haar unitary integral
+    """
+    sequence_list = [_matrix_to_sequence(matrix) for matrix in matrix_tuple]
+    row_sequence_list = [sequence[0] for sequence in sequence_list]
+    col_sequence_list = [sequence[1] for sequence in sequence_list]
+
+    def compression(sequence_list):
+        index_set = set(index for index_tuple in sequence_list for index in index_tuple)
+        index_dict = {value: index for index, value in enumerate(index_set)}
+        compressed_sequence_tuple = tuple(
+            tuple(index_dict[i] for i in index_tuple) for index_tuple in sequence_list
+        )
+
+        return len(index_set), compressed_sequence_tuple
+
+    row_count, row_compressed_seq_tuple = compression(row_sequence_list)
+    col_count, col_compressed_seq_tuple = compression(col_sequence_list)
+    matrix_list = [[[0] * col_count for _ in range(row_count)] for _ in range(2)]
+
+    for i in range(2):
+        for r, c in zip(row_compressed_seq_tuple[i], col_compressed_seq_tuple[i]):
+            matrix_list[i][r][c] += 1
 
     return [tuple(tuple(row) for row in matrix) for matrix in matrix_list]